requirements.txt
asttokens==2.2.1
backcall==0.2.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
google-auth==2.16.0
idna==3.4
ipykernel==6.20.1
ipython==8.8.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==5.1.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
pip==22.3.1
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7
pywin32==305
pyzmq==25.0.0
requests==2.28.2
rsa==4.9
setuptools==65.6.3
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
tornado==6.2
traitlets==5.8.1
urllib3==1.26.14
wcwidth==0.2.5
wheel==0.37.1
wincertstore==0.2
